# openApi/Java-面试.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Java-面试",
    "description": "",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "学生操作"
    },
    {
      "name": "班级操作"
    }
  ],
  "paths": {
    "/student/add": {
      "post": {
        "summary": " 新增学生",
        "x-apifox-folder": "学生操作",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "学生操作"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "x-apifox-ignore-properties": [],
                "x-apifox-orders": []
              },
              "example": {
                "stuNo": 123,
                "stuName": "小明",
                "stuAge": 12,
                "stuGender": 1,
                "stuClassId": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://www.apifox.cn/web/project/2564975/apis/api-73980612-run"
      }
    },
    "/student/edit": {
      "post": {
        "summary": "修改学生",
        "x-apifox-folder": "学生操作",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "学生操作"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "x-apifox-ignore-properties": [],
                "x-apifox-orders": []
              },
              "example": {
                "id": 1,
                "stuNo": 123,
                "stuName": "小明",
                "stuAge": 12,
                "stuGender": 2,
                "stuClassId": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://www.apifox.cn/web/project/2564975/apis/api-73996514-run"
      }
    },
    "/student/del/1": {
      "delete": {
        "summary": "删除学生",
        "x-apifox-folder": "学生操作",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "学生操作"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://www.apifox.cn/web/project/2564975/apis/api-73998006-run"
      }
    },
    "/student/moveClass/1/1": {
      "put": {
        "summary": "移动班级",
        "x-apifox-folder": "学生操作",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "学生操作"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://www.apifox.cn/web/project/2564975/apis/api-74002400-run"
      }
    },
    "/student/exitClass/3": {
      "put": {
        "summary": "移除班级",
        "x-apifox-folder": "学生操作",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "学生操作"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://www.apifox.cn/web/project/2564975/apis/api-74002626-run"
      }
    },
    "/class/add": {
      "post": {
        "summary": "新增班级",
        "x-apifox-folder": "班级操作",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "班级操作"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "x-apifox-ignore-properties": [],
                "x-apifox-orders": []
              },
              "example": {
                "clsName": "高一（1）班"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://www.apifox.cn/web/project/2564975/apis/api-74015463-run"
      }
    },
    "/class/edit": {
      "post": {
        "summary": "修改班级",
        "x-apifox-folder": "班级操作",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "班级操作"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "x-apifox-ignore-properties": [],
                "x-apifox-orders": []
              },
              "example": {
                "id": 1,
                "clsName": "高一（2）班"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://www.apifox.cn/web/project/2564975/apis/api-74015512-run"
      }
    },
    "/class/del/1": {
      "delete": {
        "summary": "删除班级",
        "x-apifox-folder": "班级操作",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "班级操作"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://www.apifox.cn/web/project/2564975/apis/api-74016747-run"
      }
    },
    "/student/list": {
      "get": {
        "summary": "学生列表",
        "x-apifox-folder": "学生操作",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "学生操作"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://www.apifox.cn/web/project/2564975/apis/api-74025288-run"
      }
    },
    "/class/list": {
      "get": {
        "summary": "班级列表",
        "x-apifox-folder": "班级操作",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "班级操作"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://www.apifox.cn/web/project/2564975/apis/api-74026165-run"
      }
    },
    "/class/moveClass/2/3": {
      "put": {
        "summary": "移动班级",
        "x-apifox-folder": "班级操作",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "班级操作"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://www.apifox.cn/web/project/2564975/apis/api-74027286-run"
      }
    }
  },
  "components": {
    "schemas": {}
  },
  "servers": [
    {
      "url": "http://localhost:9611",
      "description": "开发环境"
    }
  ]
}